refactor: move the repository to the repositories package
RHCLOUD-34459
